move logout function to userbutton
